Fix weird bug
Latex equations that had a symbol with subscript ending in ,pc (for example y_{tx,pc}) would cause a mtex error. Fix was to surround the pc with curly braces. It seems to be the c} characters that cause the problem.

Also added some missing [stem] lines and necessary blank lines.